+++ lib/CL/pocl_util.h
@@ -91,14 +91,14 @@ void pocl_unlock_events_inorder (cl_event ev1, cl_event ev2);
 
 /* Function for creating events */
 cl_int pocl_create_event (cl_event *event, cl_command_queue command_queue,
-                          cl_command_type command_type, int num_buffers,
+                          cl_command_type command_type, size_t num_buffers,
                           const cl_mem* buffers, cl_context context);
 
 cl_int pocl_create_command (_cl_command_node **cmd,
                             cl_command_queue command_queue,
                             cl_command_type command_type, cl_event *event,
                             cl_int num_events, const cl_event *wait_list,
-                            int num_buffers, const cl_mem *buffers);
+                            size_t num_buffers, const cl_mem *buffers);
 
 
 void pocl_command_enqueue (cl_command_queue command_queue,
